Auth route (#6)
Auth route (#6)

What changed?
=====================

This commit provides a route for the frontend to check if a user is authenticated.

Why are we doing this?
=====================

To allow the frontend to handle when a user is logged in or not. This allows users to see the page with information that only they can see and others will not be able to see.